requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1  
Django==5.2
django-cors-headers==4.7.0 
django-filter==25.1        
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
pillow==11.3.0
PyJWT==2.10.1
PyYAML==6.0.2
requests==2.32.3
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1